+++ crates/oxide_emit/src/expr/assignment.rs
@@ -11,7 +11,7 @@ use oxide_parser::AssignmentOperator;
 impl Emitter {
     pub(crate) fn emit_assignment_expression(
         &self, assign: &oxide_parser::AssignmentExpression, ctx: &mut CompileCtx,
-    ) -> Result<u8, String> {
+    ) -> Result<u32, String> {
         if let oxide_parser::AssignmentTarget::StaticMemberExpression(member) = &assign.left {
             if let Some(logical_op) = assign.operator.to_logical_operator() {
                 let store_label = ctx.next_label_id();
@@ -20,15 +20,15 @@ impl Emitter {
                 let prop_name = member.property.name.as_str();
                 let idx = ctx.add_constant(Constant::String(prop_name.to_string()));
                 let key_reg = ctx.alloc_reg();
-                ctx.inst(Inst::load_const(Operand::Reg(key_reg as u32), idx));
+                ctx.inst(Inst::load_const(Operand::Reg(key_reg), idx));
                 let result_reg = ctx.alloc_reg();
-                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg as u32), Operand::Reg(obj_reg as u32), Operand::None));
-                ctx.inst(Inst::ic_get(Operand::Reg(result_reg as u32), Operand::Reg(key_reg as u32)));
+                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg), Operand::Reg(obj_reg), Operand::None));
+                ctx.inst(Inst::ic_get(Operand::Reg(result_reg), Operand::Reg(key_reg)));
                 self.emit_logical_assign_test(logical_op, result_reg, store_label, end_label, ctx)?;
                 ctx.labels.set_label_pos(store_label, ctx.insts.len());
                 let val_reg = self.emit_expression(&assign.right, ctx)?;
-                ctx.inst(Inst::ic_set(Operand::Reg(obj_reg as u32), Operand::Reg(val_reg as u32), Operand::Reg(key_reg as u32)));
-                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg as u32), Operand::Reg(val_reg as u32), Operand::None));
+                ctx.inst(Inst::ic_set(Operand::Reg(obj_reg), Operand::Reg(val_reg), Operand::Reg(key_reg)));
+                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg), Operand::Reg(val_reg), Operand::None));
                 ctx.labels.set_label_pos(end_label, ctx.insts.len());
                 return Ok(result_reg);
             }
@@ -37,11 +37,11 @@ impl Emitter {
             let prop_name = member.property.name.as_str();
             let idx = ctx.add_constant(Constant::String(prop_name.to_string()));
             let key_reg = ctx.alloc_reg();
-            ctx.inst(Inst::load_const(Operand::Reg(key_reg as u32), idx));
+            ctx.inst(Inst::load_const(Operand::Reg(key_reg), idx));
             if assign.operator != AssignmentOperator::Assign {
-                let obj = Operand::Reg(obj_reg as u32);
-                let val = Operand::Reg(val_reg as u32);
-                let key = Operand::Reg(key_reg as u32);
+                let obj = Operand::Reg(obj_reg);
+                let val = Operand::Reg(val_reg);
+                let key = Operand::Reg(key_reg);
                 match assign.operator {
                     AssignmentOperator::Addition => ctx.inst(Inst::compound_member_add(obj, val, key)),
                     AssignmentOperator::Subtraction => ctx.inst(Inst::compound_member_sub(obj, val, key)),
@@ -53,7 +53,7 @@ impl Emitter {
                 }
                 Ok(val_reg)
             } else {
-                ctx.inst(Inst::ic_set(Operand::Reg(obj_reg as u32), Operand::Reg(val_reg as u32), Operand::Reg(key_reg as u32)));
+                ctx.inst(Inst::ic_set(Operand::Reg(obj_reg), Operand::Reg(val_reg), Operand::Reg(key_reg)));
                 Ok(val_reg)
             }
         } else if let oxide_parser::AssignmentTarget::ComputedMemberExpression(member) = &assign.left {
@@ -63,19 +63,19 @@ impl Emitter {
                 let obj_reg = self.emit_expression(&member.object, ctx)?;
                 let key_reg = self.emit_expression(&member.expression, ctx)?;
                 let result_reg = ctx.alloc_reg();
-                ctx.inst(Inst::new(OpCode::GET_PROP_DYNAMIC, Operand::Reg(obj_reg as u32), Operand::Reg(key_reg as u32), Operand::Reg(result_reg as u32)));
+                ctx.inst(Inst::new(OpCode::GET_PROP_DYNAMIC, Operand::Reg(obj_reg), Operand::Reg(key_reg), Operand::Reg(result_reg)));
                 self.emit_logical_assign_test(logical_op, result_reg, store_label, end_label, ctx)?;
                 ctx.labels.set_label_pos(store_label, ctx.insts.len());
                 let val_reg = self.emit_expression(&assign.right, ctx)?;
-                ctx.inst(Inst::new(OpCode::SET_PROP_DYNAMIC, Operand::Reg(obj_reg as u32), Operand::Reg(key_reg as u32), Operand::Reg(val_reg as u32)));
-                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg as u32), Operand::Reg(val_reg as u32), Operand::None));
+                ctx.inst(Inst::new(OpCode::SET_PROP_DYNAMIC, Operand::Reg(obj_reg), Operand::Reg(key_reg), Operand::Reg(val_reg)));
+                ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg), Operand::Reg(val_reg), Operand::None));
                 ctx.labels.set_label_pos(end_label, ctx.insts.len());
                 return Ok(result_reg);
             }
             let obj_reg = self.emit_expression(&member.object, ctx)?;
             let key_reg = self.emit_expression(&member.expression, ctx)?;
             let val_reg = self.emit_expression(&assign.right, ctx)?;
-            ctx.inst(Inst::new(OpCode::SET_PROP_DYNAMIC, Operand::Reg(obj_reg as u32), Operand::Reg(key_reg as u32), Operand::Reg(val_reg as u32)));
+            ctx.inst(Inst::new(OpCode::SET_PROP_DYNAMIC, Operand::Reg(obj_reg), Operand::Reg(key_reg), Operand::Reg(val_reg)));
             Ok(val_reg)
         } else if let oxide_parser::AssignmentTarget::PrivateFieldExpression(member) = &assign.left {
             if assign.operator != AssignmentOperator::Assign {
@@ -84,7 +84,7 @@ impl Emitter {
             let obj_reg = self.emit_expression(&member.object, ctx)?;
             let val_reg = self.emit_expression(&assign.right, ctx)?;
             let key_reg = self.emit_private_id_reg(member.field.name.as_str(), ctx)?;
-            ctx.inst(Inst::new(OpCode::SET_PRIVATE, Operand::Reg(obj_reg as u32), Operand::Reg(val_reg as u32), Operand::Reg(key_reg as u32)));
+            ctx.inst(Inst::new(OpCode::SET_PRIVATE, Operand::Reg(obj_reg), Operand::Reg(val_reg), Operand::Reg(key_reg)));
             Ok(val_reg)
         } else if let oxide_parser::AssignmentTarget::AssignmentTargetIdentifier(id_ref) = &assign.left {
             if assign.operator != AssignmentOperator::Assign {
@@ -94,14 +94,14 @@ impl Emitter {
                     let name = id_ref.name.as_str();
                     let var_reg = ctx.lookup_or_global(name);
                     let result_reg = ctx.alloc_reg();
-                    ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg as u32), Operand::Reg(var_reg as u32), Operand::None));
+                    ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg), Operand::Reg(var_reg), Operand::None));
                     self.emit_logical_assign_test(logical_op, result_reg, store_label, end_label, ctx)?;
                     ctx.labels.set_label_pos(store_label, ctx.insts.len());
                     let val_reg = self.emit_expression(&assign.right, ctx)?;
                     let is_const = ctx.lookup_const_flag(name);
                     let const_flag = if is_const { 1 } else { 0 };
-                    ctx.inst(Inst::new(OpCode::STORE_VAR, Operand::Reg(var_reg as u32), Operand::Reg(val_reg as u32), Operand::Imm(const_flag)));
-                    ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg as u32), Operand::Reg(val_reg as u32), Operand::None));
+                    ctx.inst(Inst::new(OpCode::STORE_VAR, Operand::Reg(var_reg), Operand::Reg(val_reg), Operand::Imm(const_flag)));
+                    ctx.inst(Inst::new(OpCode::LOAD_VAR, Operand::Reg(result_reg), Operand::Reg(val_reg), Operand::None));
                     ctx.labels.set_label_pos(end_label, ctx.insts.len());
                     Ok(result_reg)
                 } else if assign.operator == AssignmentOperator::Addition
@@ -135,7 +135,7 @@ impl Emitter {
                         AssignmentOperator::ShiftRightZeroFill => OpCode::COMPOUND_USHR,
                         _ => return Err(format!("compound assignment operator {:?} not supported", assign.operator)),
                     };
-                    ctx.inst(Inst::new(op, Operand::Reg(var_reg as u32), Operand::Reg(rhs as u32), Operand::None));
+                    ctx.inst(Inst::new(op, Operand::Reg(var_reg), Operand::Reg(rhs), Operand::None));
                     Ok(var_reg)
                 } else {
                     Err(format!("compound assignment operator {:?} not supported", assign.operator))
@@ -145,18 +145,18 @@ impl Emitter {
                 let name = id_ref.name.as_str();
                 // Check if target is an upvalue reference
                 if let Some(uv_idx) = ctx.current_upvalue_captures.iter().position(|u| u.name == name) {
-                    ctx.inst(Inst::new(OpCode::STORE_UPVALUE, Operand::None, Operand::Reg(val_reg as u32), Operand::Imm(uv_idx as u16)));
+                    ctx.inst(Inst::new(OpCode::STORE_UPVALUE, Operand::None, Operand::Reg(val_reg), Operand::Imm(uv_idx as u16)));
                     return Ok(val_reg);
                 }
                 // Check if target is a captured cell
                 if let Some(&cell_idx) = ctx.captured_bindings.get(name) {
-                    ctx.inst(Inst::new(OpCode::CELL_SET, Operand::None, Operand::Reg(val_reg as u32), Operand::Imm(cell_idx as u16)));
+                    ctx.inst(Inst::new(OpCode::CELL_SET, Operand::None, Operand::Reg(val_reg), Operand::Imm(cell_idx as u16)));
                     return Ok(val_reg);
                 }
                 let var_reg = ctx.lookup_or_global(name);
                 let is_const = ctx.lookup_const_flag(name);
                 let const_flag = if is_const { 1 } else { 0 };
-                ctx.inst(Inst::new(OpCode::STORE_VAR, Operand::Reg(var_reg as u32), Operand::Reg(val_reg as u32), Operand::Imm(const_flag)));
+                ctx.inst(Inst::new(OpCode::STORE_VAR, Operand::Reg(var_reg), Operand::Reg(val_reg), Operand::Imm(const_flag)));
                 Ok(val_reg)
             }
         } else if matches!(
