+++ crates/oxide_emit/src/expr/parenthesized.rs
@@ -5,7 +5,7 @@ use crate::{CompileCtx, Emitter};
 impl Emitter {
     pub(crate) fn emit_parenthesized_expression(
         &self, p: &oxide_parser::ParenthesizedExpression, ctx: &mut CompileCtx,
-    ) -> Result<u8, String> {
+    ) -> Result<u32, String> {
         self.emit_expression(&p.expression, ctx)
     }
 }
